Final V1.2: Ingredients Adapter Views Added

+++ app/src/main/java/com/example/caloriestracker/adapters/IngredientAdapter.kt
@@ -8,35 +8,68 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.caloriestracker.CreatingMealRecipe
 import com.example.caloriestracker.IngredientsActivity
 import com.example.caloriestracker.ItemModelIngredient
 import com.example.caloriestracker.R
 
-class IngredientAdapter (private val context: Context, private val list: List<ItemModelIngredient>) : RecyclerView.Adapter<IngredientAdapter.ViewHolder>() {
+class IngredientAdapter (private val context: Context, private val list: List<ItemModelIngredient>, val viewType: Int) :
+    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context)
-            .inflate(
-                R.layout.item_view_layout_ingredient,
-            parent, false)
-
-        return ViewHolder(view)
+    companion object {
+        const val VIEW_TYPE_REGULAR = 1
+        const val VIEW_TYPE_RECIPE = 2
     }
 
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val itemViewModel = list[position]
-        holder.tvName.text = itemViewModel.name
-        holder.tvCalories.text = String.format("%d kcal/100g", itemViewModel.calories)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
 
-        holder.ivModify.setOnClickListener{
-            if(context is IngredientsActivity) {
-                context.updateIngredientDialog(itemViewModel)
+        if(viewType == VIEW_TYPE_REGULAR)
+        {
+            val viewRegular = LayoutInflater.from(parent.context)
+                .inflate(
+                    R.layout.item_view_layout_ingredient,
+                    parent, false)
+
+            return ViewHolder(viewRegular)
+        } else {
+            val viewRecipe = LayoutInflater.from(parent.context)
+                .inflate(
+                    R.layout.item_view_layout_ingredient_recipe,
+                    parent, false)
+
+            return ViewHolderRecipe(viewRecipe)
+        }
+    }
+
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+
+        val itemViewModel = list[position]
+        if(holder is ViewHolder)
+        {
+            holder.tvName.text = itemViewModel.name
+            holder.tvCalories.text = String.format("%d kcal/100g", itemViewModel.calories)
+
+            holder.ivModify.setOnClickListener{
+                if(context is IngredientsActivity) {
+                    context.updateIngredientDialog(itemViewModel)
+                }
+            }
+
+            holder.ivDelete.setOnClickListener{
+                if(context is IngredientsActivity) {
+                    context.deleteIngredientDialog(itemViewModel)
+                }
             }
         }
 
-        holder.ivDelete.setOnClickListener{
-            if(context is IngredientsActivity) {
-                context.deleteIngredientDialog(itemViewModel)
+        if(holder is ViewHolderRecipe){
+            holder.tvName.text = itemViewModel.name
+            holder.tvCalories.text = String.format("%d kcal", itemViewModel.calories)
+
+            holder.ivDelete.setOnClickListener{
+                if(context is CreatingMealRecipe) {
+                    context.deleteIngredientDialog(itemViewModel)
+                }
             }
         }
     }
@@ -45,10 +78,20 @@ class IngredientAdapter (private val context: Context, private val list: List<It
         return list.size
     }
 
+    override fun getItemViewType(position: Int): Int {
+        return viewType
+    }
+
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val tvName: TextView = view.findViewById(R.id.tvName)
         val tvCalories: TextView = view.findViewById(R.id.tvCalories)
         val ivModify: ImageView = view.findViewById(R.id.ivModify)
+        val ivDelete: ImageView = view.findViewById(R.id.ivDelete)
+    }
+
+    class ViewHolderRecipe(view: View) : RecyclerView.ViewHolder(view) {
+        val tvName: TextView = view.findViewById(R.id.tvName)
+        val tvCalories: TextView = view.findViewById(R.id.tvCalories)
         val ivDelete: ImageView = view.findViewById(R.id.ivDelete)
     }
 }
